Rework Factory code to be more logical
loadTemplate() should not load plugins and addPlugins() should not instanciate DeviceRenderer: better do this in the setupRenderer() function.

loadTemplate should be launched before instance creation since are using the dom to link HTML elements to DeviceRenderer object.
Add a comment about this so that next dev will understand why it should not be reordered.