Always translate home into meta+enter
Meta+enter is always a valid shortcut to the physical home button,
so there is no need for differentiating between desktop+saas and paas.

Note the proper fix would be sending HOMEPAGE and map it to HOME,
but this would require adding the mapping to webrtcd, so we went the
easier route instead.